+++ ui/round/src/crazy/crazyView.ts
@@ -3,13 +3,13 @@ import { onInsert, LooseVNode } from 'common/snabbdom';
 import { h } from 'snabbdom';
 import RoundController from '../ctrl';
 import { Position } from '../interfaces';
-import * as round from '../round';
+import { plyStep } from '../util';
 import { crazyKeys, drag, pieceRoles } from './crazyCtrl';
 
 const eventNames = ['mousedown', 'touchstart'];
 
 export default function pocket(ctrl: RoundController, color: Color, position: Position): LooseVNode {
-  const step = round.plyStep(ctrl.data, ctrl.ply);
+  const step = plyStep(ctrl.data, ctrl.ply);
   if (!step.crazy) return;
   const droppedRole = ctrl.justDropped,
     preDropRole = ctrl.preDrop,
