+++ ui/round/src/ctrl.ts
@@ -1,7 +1,6 @@
 /// <reference types="../types/ab" />
 
 import * as ab from 'ab';
-import * as round from './round';
 import * as game from 'game';
 import { game as gameRoute } from 'game/router';
 import * as status from 'game/status';
@@ -101,11 +100,11 @@ export default class RoundController implements MoveRootCtrl {
     readonly redraw: Redraw,
     readonly nvui?: NvuiPlugin | undefined,
   ) {
-    round.massage(opts.data);
+    util.massage(opts.data);
 
     const d = (this.data = opts.data);
 
-    this.ply = round.lastPly(d);
+    this.ply = util.lastPly(d);
     this.goneBerserk[d.player.color] = d.player.berserk;
     this.goneBerserk[d.opponent.color] = d.opponent.berserk;
     setTimeout(() => {
@@ -221,7 +220,7 @@ export default class RoundController implements MoveRootCtrl {
     return true;
   };
 
-  lastPly = (): number => round.lastPly(this.data);
+  lastPly = (): number => util.lastPly(this.data);
 
   makeCgHooks = (): any => ({
     onUserMove: this.onUserMove,
@@ -247,7 +246,7 @@ export default class RoundController implements MoveRootCtrl {
   isPlaying = (): boolean => game.isPlayerPlaying(this.data);
 
   jump = (ply: Ply): boolean => {
-    ply = Math.max(round.firstPly(this.data), Math.min(this.lastPly(), ply));
+    ply = Math.max(util.firstPly(this.data), Math.min(this.lastPly(), ply));
     const isForwardStep = ply === this.ply + 1;
     this.ply = ply;
     this.justDropped = undefined;
@@ -516,7 +515,7 @@ export default class RoundController implements MoveRootCtrl {
 
   reload = (d: RoundData): void => {
     if (d.steps.length !== this.data.steps.length) this.ply = d.steps[d.steps.length - 1].ply;
-    round.massage(d);
+    util.massage(d);
     this.data = d;
     this.clearJust();
     this.shouldSendMoveTime = false;
@@ -845,7 +844,7 @@ export default class RoundController implements MoveRootCtrl {
     });
   };
 
-  stepAt = (ply: Ply): Step => round.plyStep(this.data, ply);
+  stepAt = (ply: Ply): Step => util.plyStep(this.data, ply);
 
   speakClock = (): void => {
     this.clock?.speak();
