+++ ui/round/src/ground.ts
@@ -1,10 +1,9 @@
-import * as round from './round';
 import * as util from './util';
 import resizeHandle from 'common/resize';
 import RoundController from './ctrl';
 import { Config } from 'chessground/config';
 import { h, VNode } from 'snabbdom';
-import { plyStep } from './round';
+import { plyStep } from './util';
 import { RoundData } from './interfaces';
 import { uciToMove } from 'chessground/util';
 import * as Prefs from 'common/prefs';
@@ -33,7 +32,7 @@ export function makeConfig(ctrl: RoundController): Config {
       move: hooks.onMove,
       dropNewPiece: hooks.onNewPiece,
       insert(elements) {
-        const firstPly = round.firstPly(ctrl.data);
+        const firstPly = util.firstPly(ctrl.data);
         const isSecond = (firstPly % 2 === 0 ? 'white' : 'black') !== data.player.color;
         const showUntil = firstPly + 2 + +isSecond;
         resizeHandle(
